ft_transendence/requirements.txt
Django>=3.2.3,<3.3
django-cors-headers==4.3.1
django-extensions==3.2.3
django-friendship==1.9.6
djangorestframework==3.14.0
psycopg2>=2.8.6,<2.9
uWSGI>=2.0.19.1,<2.0.24
pillow==10.2.0
djangorestframework-simplejwt==5.3.1
daphne==4.1.0
channels==4.0.0
websockets==12.0
jsonpickle==3.0.3
requests==2.31.0
pyjwt==2.8.0